pyproject.toml
[tool.poetry]
name = "llmagent"
version = "0.1.0"
description = ""
authors = ["J3ne4ek <5656tetetre@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
langchain-openai = "^0.2.4"
langchain-community = "^0.3.3"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
langchain-experimental = "^0.3.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
pylint = "^3.3.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
